requirements.txt
boto3
Flask
python-dotenv
flask_restful
requests